+++ examples/cc_vep.py
@@ -8,7 +8,7 @@ from tvb_epilepsy.base.constants.configurations import FOLDER_VEP, FOLDER_RES
 from tvb_epilepsy.base.constants.model_constants import X0_DEF, E_DEF
 from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
 from tvb_epilepsy.base.utils.log_error_utils import initialize_logger
-from tvb_epilepsy.io.writer_custom import CustomH5Writer
+from tvb_epilepsy.io.h5_writer import H5Writer
 from tvb_epilepsy.service.lsa_service import LSAService
 from tvb_epilepsy.service.model_configuration_service import ModelConfigurationService
 from tvb_epilepsy.scripts.pse_scripts import pse_from_lsa_hypothesis
@@ -17,7 +17,7 @@ from tvb_epilepsy.scripts.simulation_scripts import from_model_configuration_to_
 if DATA_MODE is TVB:
     from tvb_epilepsy.io.tvb_data_reader import TVBReader as Reader
 else:
-    from tvb_epilepsy.io.reader_custom import CustomH5Reader as Reader
+    from tvb_epilepsy.io.h5_reader import H5Reader as Reader
 
 PSE_FLAG = False
 SIM_FLAG = True
@@ -48,7 +48,7 @@ def main_vep(subject="TVB3", ep_name="clinical_hypothesis", x0_indices=[], folde
     # -------------------------------Reading data-----------------------------------
     data_folder = os.path.join(subject_folder, HEAD)
     reader = Reader()
-    writer = CustomH5Writer()
+    writer = H5Writer()
     logger.info("Reading from: " + data_folder)
     head = reader.read_head(data_folder)
     # head.plot(figure_dir=FOLDER_FIGS)
@@ -112,7 +112,6 @@ def main_vep(subject="TVB3", ep_name="clinical_hypothesis", x0_indices=[], folde
             if not (os.path.isdir(folder)):
                 os.mkdir(folder)
         logger.info("\n\nRunning hypothesis: " + hyp.name)
-        # hyp.write_to_h5(FOLDER_RES, hyp.name + ".h5")
         logger.info("\n\nCreating model configuration...")
         model_configuration_service = ModelConfigurationService(hyp.number_of_regions)
         writer.write_model_configuration_service(model_configuration_service,
